pyproject.toml
[project]
name = "basic-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "claude-agent-sdk>=0.1.5",
    "openai>=2.6.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
